PDCT-716 delete test_schemas, data migrations, ... and adapt
