+++ app/core/organisation.py
@@ -4,7 +4,7 @@ from app.api.api_v1.schemas.metadata import TaxonomyData
 from db_client.models.app.users import Organisation
 from db_client.models.law_policy.family import FamilyEventType
 from db_client.models.law_policy.metadata import MetadataOrganisation, MetadataTaxonomy
-from app.core.ingestion.types import Taxonomy, TaxonomyEntry
+from db_client.models.law_policy.taxonomy_entry import Taxonomy, TaxonomyEntry
 
 
 def get_organisation_taxonomy(db: Session, org_id: int) -> tuple[int, Taxonomy]:
