--- app/data_migrations/__init__.py
@@ -1,10 +1,0 @@
-# Pre-populate all fixed metadata tables
-from .populate_document_type import populate_document_type
-from .populate_document_role import populate_document_role
-from .populate_document_variant import populate_document_variant
-from .populate_event_type import populate_event_type
-from .populate_geo_statistics import populate_geo_statistics
-from .populate_geography import populate_geography
-from .populate_language import populate_language
-from .populate_taxonomy import populate_taxonomy
-from .populate_counters import populate_counters
