--- app/data_migrations/data/cclw/framework_data.json
@@ -1,14 +1,0 @@
-[
-  {
-    "name": "Adaptation",
-    "description": "Adaptation"
-  },
-  {
-    "name": "Mitigation",
-    "description": "Mitigation"
-  },
-  {
-    "name": "Drm/Drr",
-    "description": "Drm/Drr"
-  }
-]
