--- app/data_migrations/data/cclw/instrument_data.json
@@ -1,202 +1,0 @@
-[
-  {
-    "node": {
-      "name": "Processes, plans and strategies|Governance",
-      "description": "Processes, plans and strategies|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Subsidies|Economic",
-      "description": "Subsidies|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "International cooperation|Governance",
-      "description": "International cooperation|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Education, training and knowledge dissemination|Information",
-      "description": "Education, training and knowledge dissemination|Information",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Research & Development, knowledge generation|Information",
-      "description": "Research & Development, knowledge generation|Information",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Standards, obligations and norms|Regulation",
-      "description": "Standards, obligations and norms|Regulation",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Capacity building|Governance",
-      "description": "Capacity building|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Institutional mandates|Governance",
-      "description": "Institutional mandates|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "MRV|Governance",
-      "description": "MRV|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Tax incentives|Economic",
-      "description": "Tax incentives|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Provision of climate funds|Direct Investment",
-      "description": "Provision of climate funds|Direct Investment",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Other|Direct Investment",
-      "description": "Other|Direct Investment",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Early warning systems|Direct Investment",
-      "description": "Early warning systems|Direct Investment",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Subnational and citizen participation|Governance",
-      "description": "Subnational and citizen participation|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Insurance|Economic",
-      "description": "Insurance|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Climate finance tools|Economic",
-      "description": "Climate finance tools|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Zoning & Spatial Planning|Regulation",
-      "description": "Zoning & Spatial Planning|Regulation",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Nature based solutions and ecosystem restoration|Direct Investment",
-      "description": "Nature based solutions and ecosystem restoration|Direct Investment",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Disclosure obligations|Regulation",
-      "description": "Disclosure obligations|Regulation",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Other|Governance",
-      "description": "Other|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Planning|Governance",
-      "description": "Planning|Governance",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Green procurement|Direct Investment",
-      "description": "Green procurement|Direct Investment",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Moratoria & bans|Regulation",
-      "description": "Moratoria & bans|Regulation",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Carbon pricing & emissions trading|Economic",
-      "description": "Carbon pricing & emissions trading|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  },
-  {
-    "node": {
-      "name": "Other|Economic",
-      "description": "Other|Economic",
-      "source": "CCLW"
-    },
-    "children": []
-  }
-]
