--- app/data_migrations/data/cclw/topic_data.json
@@ -1,18 +1,0 @@
-[
-  {
-    "name": "Mitigation",
-    "description": "Mitigation"
-  },
-  {
-    "name": "Adaptation",
-    "description": "Adaptation"
-  },
-  {
-    "name": "Loss And Damage",
-    "description": "Loss And Damage"
-  },
-  {
-    "name": "Disaster Risk Management",
-    "description": "Disaster Risk Management"
-  }
-]
