--- app/data_migrations/data/law_policy/document_role_data.json
@@ -1,42 +1,0 @@
-[
-  {
-    "name": "MAIN",
-    "description": "MAIN"
-  },
-  {
-    "name": "AMENDMENT",
-    "description": "AMENDMENT"
-  },
-  {
-    "name": "SUPPORTING LEGISLATION",
-    "description": "SUPPORTING LEGISLATION"
-  },
-  {
-    "name": "SUMMARY",
-    "description": "SUMMARY"
-  },
-  {
-    "name": "PREVIOUS VERSION",
-    "description": "PREVIOUS VERSION"
-  },
-  {
-    "name": "ANNEX",
-    "description": "ANNEX"
-  },
-  {
-    "name": "SUPPORTING DOCUMENTATION",
-    "description": "SUPPORTING DOCUMENTATION"
-  },
-  {
-    "name": "INFORMATION WEBPAGE",
-    "description": "INFORMATION WEBPAGE"
-  },
-  {
-    "name": "PRESS RELEASE",
-    "description": "PRESS RELEASE"
-  },
-  {
-    "name": "DOCUMENT(S) STORED ON WEBPAGE",
-    "description": "DOCUMENT(S) STORED ON WEBPAGE"
-  }
-]
