--- app/data_migrations/data/law_policy/document_type_data.json
@@ -1,162 +1,0 @@
-[
-  {
-    "name": "Accord",
-    "description": "Accord"
-  },
-  {
-    "name": "Act",
-    "description": "Act"
-  },
-  {
-    "name": "Action Plan",
-    "description": "Action Plan"
-  },
-  {
-    "name": "Agenda",
-    "description": "Agenda"
-  },
-  {
-    "name": "Annex",
-    "description": "Annex"
-  },
-  {
-    "name": "Assessment",
-    "description": "Assessment"
-  },
-  {
-    "name": "Bill",
-    "description": "Bill"
-  },
-  {
-    "name": "Constitution",
-    "description": "Constitution"
-  },
-  {
-    "name": "Criteria",
-    "description": "Criteria"
-  },
-  {
-    "name": "Decision",
-    "description": "Decision"
-  },
-  {
-    "name": "Decision and Plan",
-    "description": "Decision and Plan"
-  },
-  {
-    "name": "Decree",
-    "description": "Decree"
-  },
-  {
-    "name": "Decree Law",
-    "description": "Decree Law"
-  },
-  {
-    "name": "Directive",
-    "description": "Directive"
-  },
-  {
-    "name": "Discussion Paper",
-    "description": "Discussion Paper"
-  },
-  {
-    "name": "Edict",
-    "description": "Edict"
-  },
-  {
-    "name": "EU Decision",
-    "description": "EU Decision"
-  },
-  {
-    "name": "EU Directive",
-    "description": "EU Directive"
-  },
-  {
-    "name": "EU Regulation",
-    "description": "EU Regulation"
-  },
-  {
-    "name": "Executive Order",
-    "description": "Executive Order"
-  },
-  {
-    "name": "Framework",
-    "description": "Framework"
-  },
-  {
-    "name": "Law",
-    "description": "Law"
-  },
-  {
-    "name": "Law and Plan",
-    "description": "Law and Plan"
-  },
-  {
-    "name": "Order",
-    "description": "Order"
-  },
-  {
-    "name": "Ordinance",
-    "description": "Ordinance"
-  },
-  {
-    "name": "Plan",
-    "description": "Plan"
-  },
-  {
-    "name": "Policy",
-    "description": "Policy"
-  },
-  {
-    "name": "Press Release",
-    "description": "Press Release"
-  },
-  {
-    "name": "Programme",
-    "description": "Programme"
-  },
-  {
-    "name": "Protocol",
-    "description": "Protocol"
-  },
-  {
-    "name": "Roadmap",
-    "description": "Roadmap"
-  },
-  {
-    "name": "Regulation",
-    "description": "Regulation"
-  },
-  {
-    "name": "Resolution",
-    "description": "Resolution"
-  },
-  {
-    "name": "Royal Decree",
-    "description": "Royal Decree"
-  },
-  {
-    "name": "Rules",
-    "description": "Rules"
-  },
-  {
-    "name": "Statement",
-    "description": "Statement"
-  },
-  {
-    "name": "Strategic Assessment",
-    "description": "Strategic Assessment"
-  },
-  {
-    "name": "Strategy",
-    "description": "Strategy"
-  },
-  {
-    "name": "Summary",
-    "description": "Summary"
-  },
-  {
-    "name": "Vision",
-    "description": "Vision"
-  }
-]
