--- app/data_migrations/data/law_policy/document_variant_data.json
@@ -1,10 +1,0 @@
-[
-  {
-    "variant_name": "Original Language",
-    "description": "Original Language"
-  },
-  {
-    "variant_name": "Translation",
-    "description": "Translation"
-  }
-]
