--- app/data_migrations/data/law_policy/event_type_data.json
@@ -1,70 +1,0 @@
-[
-  {
-    "name": "Amended",
-    "description": "Amended Event"
-  },
-  {
-    "name": "Appealed",
-    "description": "Appealed Event"
-  },
-  {
-    "name": "Closed",
-    "description": "Closed Event"
-  },
-  {
-    "name": "Declaration Of Climate Emergency",
-    "description": "Declaration Of Climate Emergency Event"
-  },
-  {
-    "name": "Dismissed",
-    "description": "Dismissed Event"
-  },
-  {
-    "name": "Entered Into Force",
-    "description": "Entered Into Force Event"
-  },
-  {
-    "name": "Filing",
-    "description": "Filing Event"
-  },
-  {
-    "name": "Granted",
-    "description": "Granted Event"
-  },
-  {
-    "name": "Implementation Details",
-    "description": "Implementation Details Event"
-  },
-  {
-    "name": "International Agreement",
-    "description": "International Agreement Event"
-  },
-  {
-    "name": "Net Zero Pledge",
-    "description": "Net Zero Pledge Event"
-  },
-  {
-    "name": "Other",
-    "description": "Other Event"
-  },
-  {
-    "name": "Passed/Approved",
-    "description": "Passed/Approved Event"
-  },
-  {
-    "name": "Repealed/Replaced",
-    "description": "Repealed/Replaced Event"
-  },
-  {
-    "name": "Set",
-    "description": "Set Event"
-  },
-  {
-    "name": "Settled",
-    "description": "Settled Event"
-  },
-  {
-    "name": "Updated",
-    "description": "Updated Event"
-  }
-]
