--- app/data_migrations/data/unf3c/submission_type_data.json
@@ -1,113 +1,0 @@
-[
-    {
-        "name": "Biennial Report"
-    },
-    {
-        "name": "Biennial Update Report"
-    },
-    {
-        "name": "Facilitative Sharing of Views Report"
-    },
-    {
-        "name": "Global Stocktake Synthesis Report"
-    },
-    {
-        "name": "Industry Report"
-    },
-    {
-        "name": "Intersessional Document"
-    },
-    {
-        "name": "Long-Term Low-Emission Development Strategy"
-    },
-    {
-        "name": "National Communication"
-    },
-    {
-        "name": "National Inventory Report"
-    },
-    {
-        "name": "Pre-Session Document"
-    },
-    {
-        "name": "Progress Report"
-    },
-    {
-        "name": "Publication"
-    },
-    {
-        "name": "Report"
-    },
-    {
-        "name": "Statement"
-    },
-    {
-        "name": "Submission to the Global Stocktake"
-    },
-    {
-        "name": "Summary Report"
-    },
-    {
-        "name": "Synthesis Report"
-    },
-    {
-        "name": "Technical Analysis Summary Report"
-    },
-    {
-        "name": "Nationally Determined Contribution"
-    },
-    {
-        "name": "Adaptation Communication"
-    },
-    {
-        "name": "National Adaptation Plan"
-    },
-    {
-        "name": "Technology Needs Assessment"
-    },
-    {
-        "name": "Fast-Start Finance Report"
-    },
-    {
-        "name": "IPCC Report"
-    },
-    {
-        "name": "Annual Compilation and Accounting Report"
-    },
-    {
-        "name": "Biennial Report,National Communication"
-    },
-    {
-        "name": "Biennial Update Report,National Communication"
-    },
-    {
-        "name": "National Adaptation Plan,Adaptation Communication"
-    },
-    {
-        "name": "National Communication,Biennial Report"
-    },
-    {
-        "name": "National Communication,Biennial Update Report"
-    },
-    {
-        "name": "Nationally Determined Contribution,Adaptation Communication"
-    },
-    {
-        "name": "Nationally Determined Contribution,National Communication"
-    },
-    {
-        "name": "Pre-Session Document,Annual Compilation and Accounting Report"
-    },
-    {
-        "name": "Pre-Session Document,Progress Report"
-    },
-    {
-        "name": "Pre-Session Document,Synthesis Report"
-    },
-    {
-        "name": "Publication,Report"
-    },
-    {
-        "name": "Technical Analysis Technical Report"
-    }
-]
