+++ tests/core/ingestion/helpers.py
@@ -1,6 +1,6 @@
 from sqlalchemy.orm import Session
 
-from app.data_migrations import (
+from db_client.data_migrations import (
     populate_document_role,
     populate_document_variant,
     populate_document_type,
