+++ tests/routes/document_helpers.py
@@ -2,7 +2,7 @@ from sqlalchemy.orm import Session
 
 
 from db_client.models.app import ORGANISATION_CCLW
-from app.data_migrations import (
+from db_client.data_migrations import (
     populate_document_role,
     populate_document_type,
     populate_document_variant,
