+++ tests/routes/setup_search_tests.py
@@ -6,8 +6,8 @@ from typing import Iterable, Mapping, Optional, Sequence
 
 from sqlalchemy.orm import Session
 
-from app.data_migrations.taxonomy_cclw import get_cclw_taxonomy
-from app.data_migrations.taxonomy_unf3c import get_unf3c_taxonomy
+from db_client.data_migrations.taxonomy_cclw import get_cclw_taxonomy
+from db_client.data_migrations.taxonomy_unf3c import get_unf3c_taxonomy
 
 from db_client.models.app import Organisation
 from db_client.models.law_policy.family import (
@@ -31,7 +31,7 @@ from db_client.models.document.physical_document import (
     PhysicalDocument,
     PhysicalDocumentLanguage,
 )
-from app.initial_data import run_data_migrations
+from db_client.initial_data import run_data_migrations
 
 FIXTURE_DIR = Path(__file__).parents[1] / "search_fixtures"
 VESPA_FAMILY_PATH = FIXTURE_DIR / "vespa_family_document.json"
