+++ tests/routes/test_config.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock
 import pytest
 
 from app.core.util import tree_table_to_json
-from app.data_migrations import (
+from db_client.data_migrations import (
     populate_document_role,
     populate_document_type,
     populate_document_variant,
